modified according to lucian's review

+++ bin/commander.py
@@ -40,36 +40,30 @@ import misc
 import vmcheckerpaths
 
 
-def main(dir):
-    # reads assignment config
-    with open(join(dir, 'config')) as handle:
-        config = ConfigParser.RawConfigParser()
-        config.readfp(handle)
+_FILES_TO_SEND = (
+    'job_build',
+    'job_run',
+    'job_errors',
+    'job_results',
+    'job_km', )
 
-    # reads vmchecker_storer.ini
-    with open(join(dir, 'storer')) as handle:
-        storer = ConfigParser.RawConfigParser()
-        storer.readfp(handle)
 
-    assignment = config.get('Assignment', 'Assignment')
-    machine = storer.get(assignment, 'Machine')
+def _run_callback(dir, ejobs):
+    """Runs callback script to upload results"""
 
-    # copies files to where vmchecker expects them (wtf
-    # XXX path is hardcoded
+    args = (join(dir, 'callback'), join(dir, 'config'))
+    args += tuple(join(ejobs, f) for f in _FILES_TO_SEND)
 
-    ejobs = vmcheckerpaths.abspath('executor_jobs')
-    # cleans up executor_jobs, if not already clean
-    if isdir(ejobs):
-        shutil.rmtree(ejobs)
-    os.mkdir(ejobs)
+    logging.info('Homework evaluated; sending results')
+    logging.debug('calling %s', args)
 
-    shutil.copy(        # copies assignment
-        join(dir, 'archive.zip'),
-        vmcheckerpaths.abspath('executor_jobs', 'file.zip'))
-    shutil.copy(        # copies tests
-        join(dir, 'tests.zip'),
-        vmcheckerpaths.abspath('executor_jobs', 'tests.zip'))
+    try:
+        check_call(args)
+    except:
+        logging.error('Sending results failed')
+        raise
 
+def _run_executor(machine, assignment):
     # starts job
     # XXX lots of wtf per minute
     # parsing config should be executors' job
@@ -91,38 +85,74 @@ def main(dir):
             ]
     logging.info('Begin homework evaluation')
     logging.debug('calling %s', args)
+
     try:
         check_call(args)
     except:
-        logging.exception('failed to run VMExecutor')
-        shutil.rmtree(ejobs)
+        logging.error('failed to run VMExecutor')
         raise
 
-    # uploads results
-    args = (
-            join(dir, 'callback'),
-            join(dir, 'config'),
-            join(ejobs, 'job_build'),
-            join(ejobs, 'job_run'),
-            join(ejobs, 'job_errors'),
-            join(ejobs, 'job_results'),
-            join(ejobs, 'job_km'),
-        )
-    logging.info('Homework evaluated; sending results')
-    logging.debug('calling %s', args)
+
+def main(dir):
+    # reads assignment config
+    with open(join(dir, 'config')) as handle:
+        config = ConfigParser.RawConfigParser()
+        config.readfp(handle)
+
+    # reads vmchecker_storer.ini
+    with open(join(dir, 'storer')) as handle:
+        storer = ConfigParser.RawConfigParser()
+        storer.readfp(handle)
+
+    # copies files to where vmchecker expects them (wtf
+    # XXX 'executor_jobs' path is hardcoded in executor
+
+    ejobs = vmcheckerpaths.abspath('executor_jobs')
+    # cleans up executor_jobs, if not already clean
+    if isdir(ejobs):
+        shutil.rmtree(ejobs)
+    os.mkdir(ejobs)
+
+    shutil.copy(        # copies assignment
+        join(dir, 'archive.zip'),
+        vmcheckerpaths.abspath('executor_jobs', 'file.zip'))
+    shutil.copy(        # copies tests
+        join(dir, 'tests.zip'),
+        vmcheckerpaths.abspath('executor_jobs', 'tests.zip'))
+
     try:
-        check_call(args)
-    except:
-        logging.exception('Sending results failed')
-        shutil.rmtree(ejobs)
-        raise
+        assignment = config.get('Assignment', 'Assignment')
+        machine = storer.get(assignment, 'Machine')
 
-    # clears files
-    shutil.rmtree(ejobs)
-    logging.info('all done')
+        _run_executor(machine, assignment)
+        _run_callback(dir, ejobs)
+
+        logging.info('all done')
+    except:
+        logging.exception('failed miserable')
+        raise
+    finally:
+        # clears files
+        shutil.rmtree(ejobs)
+
+
+def _print_help():
+    print >>sys.stderr, """Usage:
+    ./commander.py directory - where directory contains (see submit.py)
+        `homework.zip' `tests.zip' `config' `storer' `callback'"""
 
 
 if __name__ == '__main__':
     logging.basicConfig(level=logging.DEBUG)
-    assert len(sys.argv) == 2
+
+    if len(sys.argv) != 2:
+        print >>sys.stderr, 'Invalid number of arguments.'
+        _print_help()
+        exit(1)
+
+    if not os.path.isdir(sys.argv[1]):
+        print >>sys.stderr, 'Not a directory', sys.argv[1]
+        _print_help()
+        exit(1)
+
     main(sys.argv[1])
